+++ Test/Unit/Mock/HelperContextMock.php
@@ -4,7 +4,7 @@
  *
  * @package     Yireo_CheckoutTester2
  * @author      Yireo (https://www.yireo.com/)
- * @copyright   Copyright 2017 Yireo (https://www.yireo.com/)
+ * @copyright   Copyright 2022 Yireo (https://www.yireo.com/)
  * @license     Open Source License (OSL v3)
  */